requirements.txt
suds_jurko>=0.4
pyyaml>=3.1
